Handle BlockHeader events from the Ethereum blockchain
We need to keep of the current time according to the Ethereum blockchain
in order to calculate the total interest of open position, so that we
can check them against our liquidation rules.

In order to do that with the best accuracy, we maintain a persistent
connection to an Ethereum node where we subscribe for block header
messages, i.e. we will receive a block header message each time a new
block is mined, containing information on the last block, including its
timestamp.